refactor(cockpit): extract SettingsData into sibling module (#1354)
Seventh wedge of the cockpit_ui.py god-module split. Moves the
SettingsData snapshot container (data-only, __slots__-defined) into
a new cockpit_settings_data.py sibling module and re-exports it
from cockpit_ui for back-compat.

- cockpit_ui.py: 18651 -> 18615 LOC (-36)
- in-file diff: 3 inserts / 39 deletes (well under 200-line budget)
- new cockpit_settings_data.py: 64 LOC (incl. module docstring)
- no behavior change; SettingsData is constructed only in cockpit_ui
  and has zero external deps

Refs #1354.
